# Backend/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Gestion Economica e Inventario del Hogar API
  version: 1.0.0
  description: API RESTful para gestionar usuarios, casas, familias, inventario, gastos y notificaciones.
servers:
  - url: https://api.tuhogar.app/v1
    description: Servidor de produccion
  - url: http://localhost:8080/v1
    description: Servidor de desarrollo local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MembershipDto:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        family_id:
          type: integer
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
            - VIEWER
        joined_at:
          type: string
          format: date-time
...
    InventoryItemDto:
      type: object
      properties:
        id:
          type: integer
        family_id:
          type: integer
        user_id:
          type: integer
        unit_id:
          type: integer
        nombre:
          type: string
        quantity:
          type: number
        min_threshold:
          type: number
        purchase_date:
          type: string
          format: date
        expiry_date:
          type: string
          format: date
        barcode:
          type: string
...
    ReportSummaryDto:
      type: object
      properties:
        total_expenses:
          type: number
        total_income:
          type: number
        by_category:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              amount:
                type: number
    ForecastDto:
      type: object
      properties:
        projected_expenses:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              amount:
                type: number
        advice:
          type: string
    NotificationSettingDto:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        event_type:
          type: string
          enum:
            - LOW_STOCK
            - RECURRENT_EXPENSE_DUE
            - ANOMALY_DETECTED
            - EXPENSE_REMINDER
        channels:
          type: array
          items:
            type: string
            enum:
              - EMAIL
              - PUSH
              - SMS
        frequencies:
          type: array
          items:
            type: string
            enum:
              - INSTANT
              - DAILY_SUMMARY
              - WEEKLY_SUMMARY
paths:
...
  /families/{family_id}/members:
    get:
      tags:
        - House
      security:
        - bearerAuth: [ ]
      summary: Listar membresias
      parameters:
        - in: path
          name: family_id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Lista de membresias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MembershipDto'
    post:
      tags:
        - House
      security:
        - bearerAuth: [ ]
      summary: Invitar miembro
      parameters:
        - in: path
          name: family_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - role
              properties:
                user_id:
                  type: integer
                role:
                  type: string
                  enum:
                    - ADMIN
                    - MEMBER
                    - VIEWER
      responses:
        '201':
          description: Invitacion enviada
  /families/{family_id}/members/{id}:
    put:
      tags:
        - House
      security:
        - bearerAuth: [ ]
      summary: Actualizar rol de miembro
      parameters:
        - in: path
          name: family_id
          schema:
            type: integer
          required: true
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum:
                    - ADMIN
                    - MEMBER
                    - VIEWER
      responses:
        '200':
          description: Rol actualizado
...
  /inventory-items:
    get:
      tags:
        - Inventory
      security:
        - bearerAuth: [ ]
      summary: Listar items de inventario
      parameters:
        - in: query
          name: family_id
          schema:
            type: integer
          required: true
        - in: query
          name: threshold
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItemDto'
    post:
      tags:
        - Inventory
      security:
        - bearerAuth: [ ]
      summary: Crear item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItemDto'
      responses:
        '201':
          description: item creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemDto'
  /inventory-items/{id}:
    put:
      tags:
        - Inventory
      security:
        - bearerAuth: [ ]
      summary: Actualizar item
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: number
                min_threshold:
                  type: number
                purchase_date:
                  type: string
                  format: date
                expiry_date:
                  type: string
                  format: date
      responses:
        '200':
          description: item actualizado
...
  /reports/summary:
    get:
      tags:
        - Reports
      security:
        - bearerAuth: [ ]
      summary: Resumen de ingresos y egresos
      parameters:
        - in: query
          name: family_id
          schema:
            type: integer
          required: true
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Resumen calculado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSummaryDto'
  /reports/forecast:
    get:
      tags:
        - Reports
      security:
        - bearerAuth: [ ]
      summary: Proyeccion de gastos futuros
      parameters:
        - in: query
          name: family_id
          schema:
            type: integer
          required: true
        - in: query
          name: horizon_days
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Proyeccion de gastos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastDto'
  /notification-settings:
    get:
      tags:
        - Notification
      security:
        - bearerAuth: [ ]
      summary: Listar configuraciones de notificacion
      responses:
        '200':
          description: Lista de NotificationSettingDto
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationSettingDto'
  /notification-settings/{id}:
    put:
      tags:
        - Notification
      security:
        - bearerAuth: [ ]
      summary: Actualizar configuracion de notificacion
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSettingDto'
      responses:
        '200':
          description: Configuracion de notificacion actualizada